pyproject.toml
[tool.poetry]
name = "compcheck"
version = "0.1.0"
description = ""
authors = ["zikden <rivming@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.7.2"
beautifulsoup4 = "4.12.2"
django = "4.2.9"
django-bootstrap-v5 = "1.0.11"
pillow = "10.2.0"
python-dotenv = "1.0.1"
soupsieve = "2.5"
sqlparse = "0.4.4"
tzdata = "2023.4"
djangorestframework = "^3.15.1"
requests = "^2.31.0"
redis = "^5.0.4"
celery = "^5.4.0"
django-debug-toolbar = "^4.3.0"
django-filter = "^24.2"
selenium = "^4.20.0"
lxml = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
